Adapt database code to latest drift change

+++ lib/core/datasources/saved_posts_local_data_source.dart
@@ -66,7 +66,7 @@ class SavedPostsLocalDataSource extends DatabaseAccessor<AppDatabase>
 
       return PostListModel.fromDB(
         await raw,
-        await count,
+        await count as int,
       );
     } catch (error) {
       throw DatabaseException();
